# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CleanFlow ERP API",
    "version": "0.1.0",
    "description": "REST API modular monolith for dry-cleaning operations"
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "refreshCookie": { "type": "apiKey", "in": "cookie", "name": "cleanflow_refresh" }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/nomenclature/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "patch": { "responses": { "200": { "description": "Nomenclature item updated" } } },
      "delete": {
        "responses": { "200": { "description": "Nomenclature item archived" } }
      }
    },
...
    "/defect-groups/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "patch": { "responses": { "200": { "description": "Defect group updated" } } },
      "delete": { "responses": { "200": { "description": "Defect group archived" } } }
    },
...
  }
}